move reusable functions to their own file
